rename GHEventsListener to Subscriber and move ext point to separate pkg
